init login refactor
