+++ src/applications/login/components/SignInWrapper.jsx
@@ -1,9 +1,5 @@
 import React from 'react';
 
 export default function SignInWrapper({ children }) {
-  return (
-    <section className="login">
-      <div className="container">{children}</div>
-    </section>
-  );
+  return <>{children}</>;
 }
