+++ src/platform/user/authentication/components/index.jsx
@@ -1,0 +1,4 @@
+export { default as LoginContainer } from './LoginContainer';
+export { default as LoginActions } from './LoginActions';
+export { default as LoginInfo } from './LoginInfo';
+export { default as LoginHeader } from './LoginHeader';
