+++ src/platform/user/tests/authentication/components/loginActions.unit.spec.js
@@ -4,7 +4,7 @@ import { mount } from 'enzyme';
 import sinon from 'sinon';
 
 import * as authUtilities from '../../../authentication/utilities';
-import LoginContainer from '../../../authentication/components/LoginContainer';
+import LoginActions from '../../../authentication/components/LoginActions';
 import { CSP_IDS } from '../../../authentication/constants';
 
 describe('login DOM ', () => {
@@ -19,7 +19,7 @@ describe('login DOM ', () => {
   });
 
   it('login buttons should properly call login method', () => {
-    const loginButtons = mount(<LoginContainer />);
+    const loginButtons = mount(<LoginActions />);
 
     const testButton = button => {
       const loginCSP = button.prop('data-csp');
